download remote files on import

+++ src/test/java/org/sergeys/cookbook/test/JsoupTest.java
@@ -8,7 +8,12 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.io.UnsupportedEncodingException;
+import java.net.URI;
 import java.net.URLDecoder;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+import java.net.http.HttpResponse.BodyHandlers;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystems;
@@ -63,7 +68,7 @@ class JsoupTest {
         elts.remove();
     }
 
-    private void fixReferences(Document doc, String tag, String attribute,
+    private void copyReferencedFiles(Document doc, String tag, String attribute,
             String srcMainDirname, String targetMainDirname, String targetSubdirname) {
 
         String targetDirname = targetMainDirname + File.separator + targetSubdirname;
@@ -77,10 +82,47 @@ class JsoupTest {
         elements.parallelStream().forEach(e -> {
             String ref = e.attr(attribute); // expected relative path with subdir, e.g ./subdir/file.ext
 
-            if(ref.startsWith("http:") || ref.startsWith("https:") || ref.isEmpty()) {
-                log.debug("skipping element " + tag + ", " + attribute + ": " + ref);
+//            if(ref.startsWith("http:") || ref.startsWith("https:") || ref.isEmpty()) {
+//                log.debug("skipping element " + tag + ", " + attribute + ": " + ref); //TODO download remote files?
+//                return;
+//            }
+
+            if(ref.isEmpty()) {
+                log.debug("skipping empty ref element " + tag + ", " + attribute + ": " + ref);
                 return;
             }
+
+
+            // ignore case sensitivity
+            if(ref.startsWith("http:") || ref.startsWith("https:")){
+                // remote file
+                if(
+                    ref.endsWith(".js")
+                    || ref.endsWith(".css")
+                    || ref.endsWith(".jpg")
+                    || ref.endsWith(".ico")
+                    // TODO parse supposed filename between last / and ?
+                ) {
+                    log.debug("download " + ref);
+                    HttpClient client = HttpClient.newBuilder()
+                            .build();
+                    HttpRequest req = HttpRequest.newBuilder()
+                            .GET()
+                            .uri(URI.create(ref))
+                            .build();
+                    try {
+                        HttpResponse<byte[]> resp = client.send(req, BodyHandlers.ofByteArray());
+                        log.debug("got " + resp.body().length + " bytes");
+                    } catch (IOException | InterruptedException e1) {
+                        log.error("failed", e1);
+                        return;
+                    }
+                }
+                else {
+                    log.debug("skipping element " + tag + ", " + attribute + ": " + ref + ", looks like not a file");
+                }
+            }
+
 
             // restore e.g cyrillic text encoded as %D0%BF%D0%B0%D1
             ref= URLDecoder.decode(ref, StandardCharsets.UTF_8);
@@ -94,9 +136,18 @@ class JsoupTest {
 
             // assume saved files keep other files in their relative subdir
             Path src = FileSystems.getDefault().getPath(srcMainDirname, ref);
+            if(!src.toFile().exists()) {
+                log.error("File " + src + " does not exist");
+                return;
+            }
             Path dest = FileSystems.getDefault().getPath(targetDirname, src.toFile().getName());
             try {
-                Files.copy(src, dest, StandardCopyOption.REPLACE_EXISTING);
+                if(!dest.toFile().exists()) {
+                    Files.copy(src, dest, StandardCopyOption.REPLACE_EXISTING); // still filealreadyexists exception?
+                }
+                else {
+                    //log.debug(dest + " already exists");
+                }
             } catch (IOException e1) {
                 log.error("failed", e1);
             }
@@ -115,6 +166,9 @@ class JsoupTest {
         //File inFile = new File("d:/git/cookbook/samples/Буженина в фольге - Все проходит....html");
         //File inFile = new File("d:/git/cookbook/samples/паста милано.htm");
         File inFile = new File("/home/sergeys/tmp/Суд да дело  dok_zlo — ЖЖ.html");
+        //File inFile = new File("/home/sergeys/tmp/TyqYxuUgdA3E9b.html");
+
+
         try {
             Document doc = Jsoup.parse(inFile, null);
 
@@ -122,21 +176,24 @@ class JsoupTest {
 
             // remove junk
 
+            // looks broken if leave scripts:
+            // https://ru-kitchen.ru/TyqYxuUgdA3E9b
             removeElements(doc, "script");
             removeElements(doc, "noscript");
+            removeElements(doc, "noindex");
 
 
             // adjust relative references and copy referenced files
 
-            String newName = "sample-output";
+            String newName = "sample-output2";
             String baseOutputDir = Settings.getSettingsDirPath();
 
             //String targetSubdirName = baseOutputDir + File.separator + newName;
             String targetMainFile = baseOutputDir + File.separator + newName + ".html";
 
-            fixReferences(doc, "img", "src",
+            copyReferencedFiles(doc, "img", "src",
                     inFile.getParentFile().getAbsolutePath(), baseOutputDir, newName);
-            fixReferences(doc, "link", "href",
+            copyReferencedFiles(doc, "link", "href",
                     inFile.getParentFile().getAbsolutePath(), baseOutputDir, newName);
 
             // write modified main file
